+++ zamp_sdk/action_executor/constants/__init__.py
@@ -3,6 +3,12 @@ from zamp_sdk.action_executor.constants.polling import (
     POLL_MAX_INTERVAL_SECONDS,
     POLL_TIMEOUT_SECONDS,
 )
+from zamp_sdk.action_executor.constants.retries import (
+    RETRY_5XX_BACKOFF_COEFFICIENT,
+    RETRY_5XX_INITIAL_INTERVAL_SECONDS,
+    RETRY_5XX_MAX_ATTEMPTS,
+    RETRY_5XX_MAX_INTERVAL_SECONDS,
+)
 from zamp_sdk.action_executor.constants.statuses import (
     IN_PROGRESS_STATUSES,
     SUCCESS_STATUSES,
@@ -16,6 +22,10 @@ __all__ = [
     "POLL_INITIAL_INTERVAL_SECONDS",
     "POLL_MAX_INTERVAL_SECONDS",
     "POLL_TIMEOUT_SECONDS",
+    "RETRY_5XX_BACKOFF_COEFFICIENT",
+    "RETRY_5XX_INITIAL_INTERVAL_SECONDS",
+    "RETRY_5XX_MAX_ATTEMPTS",
+    "RETRY_5XX_MAX_INTERVAL_SECONDS",
     "SUCCESS_STATUSES",
     "TERMINAL_FAILURE_STATUSES",
 ]